changed floats to int

+++ src/market.rs
@@ -3,10 +3,10 @@ use log::{info};
 pub struct Good {
     pub(crate) price: f32,
     name: String,
-    q_sold: f32,
-    q_bought: f32,
-    q_demanded: Vec<f32>,
-    q_supplied: Vec<f32>,
+    q_sold: i32,
+    q_bought: i32,
+    q_demanded: Vec<i32>,
+    q_supplied: Vec<i32>,
 }
 pub struct Market {
     goods: Vec<Good>,
@@ -16,7 +16,7 @@ impl Market {
         Market { goods: Vec::new() }
     }
     pub fn add_good(&mut self, price: f32, name: String) {
-        self.goods.push(Good { price, name , q_sold: 0.0, q_bought: 0.0, q_demanded: Vec::new(), q_supplied: Vec::new() });
+        self.goods.push(Good { price, name , q_sold: 0, q_bought: 0, q_demanded: Vec::new(), q_supplied: Vec::new() });
     }
     pub fn get_good(&self, key: &String) -> Option<&Good>{
         for good in self.goods.iter() {
@@ -31,12 +31,12 @@ impl Market {
         for good in self.goods.iter_mut() {
             good.q_demanded.push(good.q_sold );
             good.q_supplied.push(good.q_bought );
-            good.q_sold = 0.0;
-            good.q_bought  = 0.0;
+            good.q_sold = 0;
+            good.q_bought  = 0;
         }
     }
     
-    pub fn increase_q_sold(&mut self, name: String, amount: f32) {
+    pub fn increase_q_sold(&mut self, name: String, amount: i32) {
         for good in self.goods.iter_mut() {
             if good.name == name {
                 good.q_sold += amount;
@@ -44,7 +44,7 @@ impl Market {
             }
         }
     }
-    pub fn increase_q_bought(&mut self, name: String, amount: f32) {
+    pub fn increase_q_bought(&mut self, name: String, amount: i32) {
         for good in self.goods.iter_mut() {
             if good.name == name {
                 good.q_bought += amount;
@@ -55,7 +55,7 @@ impl Market {
     
     pub fn update_good_price(&mut self) {
         for good in self.goods.iter_mut() {
-            good.price =  good.price * (good.q_bought / good.q_sold).sqrt();
+            good.price =  good.price * (good.q_bought as f32 / good.q_sold as f32).sqrt();
             info!("The price of {} is now {}, Q_bought: {} , Q_Sold: {}", good.name, good.price, good.q_bought , good.q_sold);
         }
     }
